修改字典值逻辑.

+++ orion-ops-module-infra/orion-ops-module-infra-service/src/main/java/com/orion/ops/module/infra/service/impl/DictKeyServiceImpl.java
@@ -3,6 +3,7 @@ package com.orion.ops.module.infra.service.impl;
 import com.alibaba.fastjson.JSON;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.orion.lang.utils.Objects1;
+import com.orion.ops.framework.biz.operator.log.core.uitls.OperatorLogs;
 import com.orion.ops.framework.common.constant.Const;
 import com.orion.ops.framework.common.constant.ErrorMessage;
 import com.orion.ops.framework.common.utils.Valid;
@@ -15,6 +16,7 @@ import com.orion.ops.module.infra.entity.dto.DictKeyCacheDTO;
 import com.orion.ops.module.infra.entity.request.dict.DictKeyCreateRequest;
 import com.orion.ops.module.infra.entity.request.dict.DictKeyUpdateRequest;
 import com.orion.ops.module.infra.entity.vo.DictKeyVO;
+import com.orion.ops.module.infra.enums.DictValueTypeEnum;
 import com.orion.ops.module.infra.service.DictKeyService;
 import com.orion.ops.module.infra.service.DictValueService;
 import lombok.extern.slf4j.Slf4j;
@@ -45,7 +47,8 @@ public class DictKeyServiceImpl implements DictKeyService {
 
     @Override
     public Long createDictKey(DictKeyCreateRequest request) {
-        log.info("DictKeyService-createDictKey request: {}" , JSON.toJSONString(request));
+        log.info("DictKeyService-createDictKey request: {}", JSON.toJSONString(request));
+        Valid.valid(DictValueTypeEnum::of, request.getValueType());
         // 转换
         DictKeyDO record = DictKeyConvert.MAPPER.to(request);
         // 查询数据是否冲突
@@ -53,7 +56,7 @@ public class DictKeyServiceImpl implements DictKeyService {
         // 插入
         int effect = dictKeyDAO.insert(record);
         Long id = record.getId();
-        log.info("DictKeyService-createDictKey id: {}, effect: {}" , id, effect);
+        log.info("DictKeyService-createDictKey id: {}, effect: {}", id, effect);
         // 删除缓存
         RedisMaps.delete(DictCacheKeyDefine.DICT_KEY);
         return id;
@@ -63,10 +66,11 @@ public class DictKeyServiceImpl implements DictKeyService {
     @Transactional(rollbackFor = Exception.class)
     public Integer updateDictKeyById(DictKeyUpdateRequest request) {
         Long id = Valid.notNull(request.getId(), ErrorMessage.ID_MISSING);
-        log.info("DictKeyService-updateDictKeyById id: {}, request: {}" , id, JSON.toJSONString(request));
+        log.info("DictKeyService-updateDictKeyById id: {}, request: {}", id, JSON.toJSONString(request));
+        Valid.valid(DictValueTypeEnum::of, request.getValueType());
         // 查询
         DictKeyDO record = dictKeyDAO.selectById(id);
-        Valid.notNull(record, ErrorMessage.DATA_ABSENT);
+        Valid.notNull(record, ErrorMessage.CONFIG_ABSENT);
         // 转换
         DictKeyDO updateRecord = DictKeyConvert.MAPPER.to(request);
         // 查询数据是否冲突
@@ -79,7 +83,7 @@ public class DictKeyServiceImpl implements DictKeyService {
         }
         // 删除缓存
         RedisMaps.delete(DictCacheKeyDefine.DICT_KEY);
-        log.info("DictKeyService-updateDictKeyById effect: {}" , effect);
+        log.info("DictKeyService-updateDictKeyById effect: {}", effect);
         return effect;
     }
 
@@ -110,27 +114,38 @@ public class DictKeyServiceImpl implements DictKeyService {
 
     @Override
     public Integer deleteDictKeyById(Long id) {
-        log.info("DictKeyService-deleteDictKeyById id: {}" , id);
+        log.info("DictKeyService-deleteDictKeyById id: {}", id);
         // 检查数据是否存在
         DictKeyDO record = dictKeyDAO.selectById(id);
-        Valid.notNull(record, ErrorMessage.DATA_ABSENT);
+        Valid.notNull(record, ErrorMessage.CONFIG_ABSENT);
+        OperatorLogs.add(OperatorLogs.KEY, record.getKey());
         // 删除配置项
         int effect = dictKeyDAO.deleteById(id);
         // 删除配置值
         dictValueService.deleteDictValueByKeyId(id);
         // 删除缓存
         RedisMaps.delete(DictCacheKeyDefine.DICT_KEY, id);
-        log.info("DictKeyService-deleteDictKeyById id: {}, effect: {}" , id, effect);
+        log.info("DictKeyService-deleteDictKeyById id: {}, effect: {}", id, effect);
         return effect;
     }
 
     @Override
     public Integer deleteDictKeyByIdList(List<Long> idList) {
-        log.info("DictKeyService-deleteDictKeyByIdList idList: {}" , idList);
+        log.info("DictKeyService-deleteDictKeyByIdList idList: {}", idList);
+        // 检查数据是否存在
+        List<DictKeyDO> dictKeys = dictKeyDAO.selectBatchIds(idList);
+        if (dictKeys.isEmpty()) {
+            return 0;
+        }
+        String keys = dictKeys.stream()
+                .map(DictKeyDO::getKey)
+                .collect(Collectors.joining(Const.COMMA));
+        OperatorLogs.add(OperatorLogs.KEY, keys);
+        // 删除配置项
         int effect = dictKeyDAO.deleteBatchIds(idList);
         // 删除配置值
         dictValueService.deleteDictValueByKeyIdList(idList);
-        log.info("DictKeyService-deleteDictKeyByIdList effect: {}" , effect);
+        log.info("DictKeyService-deleteDictKeyByIdList effect: {}", effect);
         // 删除缓存
         RedisMaps.delete(DictCacheKeyDefine.DICT_KEY, idList);
         return effect;
